+++ R/Connection.R
@@ -29,7 +29,9 @@ AthenaConnection <-
     s3_staging_dir = NULL,
     region_name = NULL,
     profile_name = NULL, 
-    aws_expiration = NULL,...){
+    aws_expiration = NULL,
+    keyboard_interrupt = NULL,
+    ...){
     
     # get lower level paws methods
     get_region <- pkg_method("get_region", "paws.common")
@@ -67,6 +69,7 @@ AthenaConnection <-
                  dbms.name = schema_name, work_group = work_group %||% "primary",
                  poll_interval = poll_interval, encryption_option = encryption_option,
                  kms_key = kms_key, expiration = aws_expiration,
+                 keyboard_interrupt = keyboard_interrupt,
                  region_name = RegionName)
     
     res <- new("AthenaConnection",  ptr = ptr, info = info, quote = "`")