+++ R/utils.R
@@ -15,8 +15,24 @@ is.s3_uri <- function(x) {
   grepl(regex, x)
 }
 
-# holds functions until athena query competed
+
+# If Query is cancelled by keyboard interrupt stop AWS Athena process
 poll <- function(res){
+  tryCatch(.poll(res),
+           interrupt = function(i){
+             if(res@connection@info$keyboard_interrupt){
+               dbClearResult(res)
+               stop(sprintf("Query '%s' has been cancelled by user.", res@info$QueryExecutionId), call. = F)
+             }
+             else {
+               stop(sprintf("Query '%s' has been cancelled by user but will carry on running in AWS Athena", res@info$QueryExecutionId), call. = F)
+             }
+           }
+  )
+}
+
+# holds functions until athena query competed
+.poll <- function(res){
   class_poll <- res@connection@info$poll_interval
   while (TRUE){
     poll_interval <- class_poll %||% rand_poll()